feat: add raw version of ContentfulContent
